pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "face-duplicate-detector"
version = "1.0.0"
description = "GPU-Accelerated Face Duplicate Detection System"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Face Detection Team", email = "team@example.com"},
]
keywords = ["face-detection", "computer-vision", "gpu", "cuda", "machine-learning"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Video",
]
requires-python = ">=3.8"
dependencies = [
    "face-recognition>=1.3.0",
    "opencv-python>=4.8.1",
    "numpy>=1.24.3",
    "Pillow>=10.0.1",
    "torch>=2.0.1",
    "torchvision>=0.15.2",
    "facenet-pytorch>=2.5.3",
    "typer[all]>=0.9.0",
    "rich>=13.5.2",
    "loguru>=0.7.2",
    "tqdm>=4.66.1",
    "pandas>=2.0.3",
    "scikit-learn>=1.3.0",
    "moviepy>=1.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/username/face-duplicate-detector"
Repository = "https://github.com/username/face-duplicate-detector"
Documentation = "https://github.com/username/face-duplicate-detector/wiki"
"Bug Tracker" = "https://github.com/username/face-duplicate-detector/issues"

[project.scripts]
face-detector = "main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
